requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document parsing
# Using PyMuPDF4llm which has pre-built wheels for Python 3.13
PyMuPDF4llm>=0.0.5
python-docx==1.1.0
docx2txt==0.8
pdfminer.six==20221105

# NLP & Embeddings
spacy==3.7.2
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4

# LLM Integration
ollama==0.1.6
openai==1.10.0
anthropic==0.18.0  # For Claude models
google-generativeai==0.3.2  # For Gemini models

# Utilities
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
python-multipart==0.0.6
aiofiles==23.2.1

# UI
streamlit==1.30.0
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
